Added a model for the server-side representation of the Client info. This stores the information from client transmitted via hello().
Storing this information would allow selecting specific clients e.g. for purposes of sending system messages but also for purely informational (dashboard) reasons.

Currently the assumption is that one client is only ever running on one physical client. The abstraction that a TalkClient can be run simultaneously on several physical devices is not reflected.

This is probably ok this way since multi-device support may function entirely differently in the future by means of channels, where one channel is the abstract representation of what currently is TalkClient.